% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iDOM-package.R
\docType{package}
\name{iDOM-package}
\alias{iDOM}
\alias{iDOM-package}
\title{iDOM: Package for Dissolved organic matter analysis}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jianjun Wang \email{jjwang@niglas.ac.cn} (\href{https://orcid.org/0000-0001-7039-7136}{ORCID})

Authors:
\itemize{
  \item Ang Hu \email{anghu@hunau.edu.cn} (\href{https://orcid.org/0000-0002-5755-2442}{ORCID})
  \item Fanfan Meng \email{mengfanfan19@mails.ucas.ac.cn} (\href{https://orcid.org/0000-0001-5718-1621}{ORCID})
}

}
\keyword{internal}
