% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Microbial.data}
\alias{Microbial.data}
\title{Microbial community data}
\format{
The format for Microbial community data is a data.frame with samples in the rows and species in the columns.
}
\source{
\url{https://doi.org/10.1038/ncomms13960}.
}
\usage{
Microbial.data
}
\description{
Example data of a Microbial compositional table (50 samples by 100 bacterial species)
}
\examples{
data(Microbial.data)
}
\references{
Wang J, Pan F, Soininen J, Heino J, Shen J. 2016.
Nutrient enrichment modifies temperature-biodiversity relationships in large-scale field experiments. \emph{Nature Communications} 7:1-9.

Hu A, Choi M, Tanentzap AJ, Liu J, Jang K-S, Lennon JT, Liu Y, Soininen J, Lu X, Zhang Y, Shen J, Wang J. 2021.
Quantifying microbial associations of dissolved organic matter under global change. \emph{bioRxiv}.

Hu A, Choi M, Tanentzap AJ, Liu J, Jang K-S, Lennon JT, Liu Y, Soininen J, Lu X, Zhang Y, Shen J, Wang J. 2022.
Quantifying the associations between dissolved organic matter and microbes under global change. \emph{Nature communications}.
}
\keyword{datasets}
