% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DOM.data}
\alias{DOM.data}
\title{DOM molecules data}
\format{
The format for DOM data is a data.frame with samples in the rows and DOM molecules in the columns.
}
\usage{
DOM.data
}
\description{
Example data of a DOM compositional table (50 samples by 100 DOM molecules)
}
\examples{
data(DOM.data)
}
\references{


Hu A, Choi M, Tanentzap AJ, Liu J, Jang K-S, Lennon JT, et al.
Quantifying the associations between dissolved organic matter and microbes under global change. \emph{Nature communications} 2022.
}
\keyword{datasets}
