% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDTiA.H2.R
\name{EDTiA.H2}
\alias{EDTiA.H2}
\title{FUNCTION_TITLE}
\usage{
EDTiA.H2(
  Comm.Bacteria,
  Comm.DOM,
  occurrence.threshold = 0.5,
  threshold.r = 0.3,
  N = 100,
  Null.model = "swap.web"
)
}
\arguments{
\item{Comm.Bacteria}{PARAM_DESCRIPTION}

\item{Comm.DOM}{PARAM_DESCRIPTION}

\item{occurrence.threshold}{PARAM_DESCRIPTION, Default: 0.5}

\item{threshold.r}{PARAM_DESCRIPTION, Default: 0.3}

\item{N}{PARAM_DESCRIPTION, Default: 100}

\item{Null.model}{PARAM_DESCRIPTION, Default: 'swap.web'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[SpiecEasi]{sparcc}}
\code{\link[bipartite]{nullmodel}}, \code{\link[bipartite]{networklevel}}
}
