% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Microbial.data}
\alias{Microbial.data}
\title{Microbial community data}
\format{
The format for community data is a data.frame with samples in the rows and species in the columns.
}
\source{
\url{https://doi.org/10.1038/ncomms13960}.
}
\usage{
Microbial.data
}
\description{
Example data of a bacterial compositional table (50 samples by 100 bacterial species)
}
\examples{
data(Microbial.data)
}
\references{
Wang J, Pan F, Soininen J, et al. Nutrient enrichment modifies temperature-biodiversity relationships in large-scale field experiments.
\emph{Nature Communications}, 2016, 7(1): 1-9.

Hu A, Choi M, Tanentzap AJ, Liu J, Jang K-S, Lennon JT, et al.
Quantifying the associations between dissolved organic matter and microbes under global change. \emph{Nature communications} 2022.
}
\keyword{datasets}
